requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
pydantic==1.10.7
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.1.0
pillow==9.5.0
jinja2==3.1.2
python-slugify==8.0.1
